+++ Utils/D19cPSEventAS.cc
@@ -69,7 +69,7 @@ void D19cPSEventAS::Set(const BeBoard* pBoard, const std::vector<uint32_t>& pDat
     }             // optical groups
 }
 // required by event but not sure if makes sense for AS
-void D19cPSEventAS::fillChipDataContainer(ChipDataContainer* chipContainer, const std::shared_ptr<ChannelGroupBase> testChannelGroup, uint16_t hybridId)
+void D19cPSEventAS::fillChipDataContainer(ChipDataContainer* chipContainer, const std::shared_ptr<ChannelGroupBase> testChannelGroup, uint8_t hybridId)
 {
     uint8_t cHybridIndex = getHybridIndex(hybridId);
     uint8_t cChipIndex   = getChipIndex(cHybridIndex, chipContainer->getId());
